Initial doxygen tag inclusion, and split out classes into their own header files
