+++ include/NCPA/atmosphere/abstract_atmospheric_property.hpp
@@ -1,41 +1,85 @@
 #pragma once
 
+/**
+ * @file
+ * @brief
+ */
+
 #include "NCPA/atmosphere/declarations.hpp"
 
 #include <memory>
 #include <stdexcept>
 
-namespace NCPA {
-    namespace atmos {
-        class abstract_atmospheric_property;
-    }
-}  // namespace NCPA
-
-#define RETURN_THIS_AS_ABSTRACT_ATMOSPHERIC_PROPERTY \
-    return static_cast<abstract_atmospheric_property&>( *this );
-
+/**
+ * @brief
+ * @param a
+ * @param b
+ */
 static void swap( NCPA::atmos::abstract_atmospheric_property&,
                   NCPA::atmos::abstract_atmospheric_property& ) noexcept;
 
 namespace NCPA {
     namespace atmos {
+
+        /**
+         * @brief
+         */
         class abstract_atmospheric_property {
             public:
+                /**
+                 * @brief
+                 */
                 virtual ~abstract_atmospheric_property() {}
+
+                /**
+                 * @brief
+                 * @return size_t
+                 */
                 virtual size_t dimensions() const = 0;
+
+                /**
+                 * @brief
+                 * @return std::unique_ptr<abstract_atmospheric_property>
+                 */
                 virtual std::unique_ptr<abstract_atmospheric_property> clone()
-                    const
-                    = 0;
+                    const = 0;
+
+                /**
+                 * @brief
+                 * @param source
+                 * @return abstract_atmospheric_property&
+                 */
                 virtual abstract_atmospheric_property& copy(
-                    const abstract_atmospheric_property& source )
-                    = 0;
+                    const abstract_atmospheric_property& source ) = 0;
+
+                /**
+                 * @brief
+                 * @return true
+                 * @return false
+                 */
                 virtual bool strict() const { return _strict; }
+
+                /**
+                 * @brief
+                 * @param tf
+                 * @return abstract_atmospheric_property&
+                 */
                 virtual abstract_atmospheric_property& set_strict( bool tf ) {
                     _strict = tf;
                     return *this;
                 }
+
             protected:
+                /**
+                 * @brief
+                 */
                 bool _strict = false;
+
+                /**
+                 * @brief
+                 * @param n
+                 * @throws std::range_error
+                 */
                 virtual void _validate_axis( size_t n ) const {
                     if (n >= this->dimensions()) {
                         std::ostringstream oss;
@@ -44,7 +88,5 @@ namespace NCPA {
                     }
                 }
         };
-
-        typedef std::unique_ptr<abstract_atmospheric_property> _atm_prop_ptr_t;
     }  // namespace atmos
 }  // namespace NCPA
