+++ include/NCPA/atmosphere/ncpaprop_atmosphere_reader_stratified_2d.hpp
@@ -11,7 +11,6 @@
 #include "NCPA/atmosphere/Atmosphere1D.hpp"
 #include "NCPA/atmosphere/Atmosphere2D.hpp"
 #include "NCPA/atmosphere/Atmosphere3D.hpp"
-// #include "NCPA/atmosphere/builders.hpp"
 #include "NCPA/atmosphere/declarations.hpp"
 #include "NCPA/atmosphere/ncpaprop_atmosphere_reader_2d.hpp"
 #include "NCPA/files.hpp"
@@ -37,9 +36,31 @@ namespace NCPA {
                     const ncpaprop_atmosphere_reader_stratified_2d& other ) :
                     ncpaprop_atmosphere_reader_2d() {}
 
-                DECLARE_BOILERPLATE_METHODS(
-                    ncpaprop_atmosphere_reader_stratified_2d,
-                    _abstract_atmosphere_reader_2d )
+                ncpaprop_atmosphere_reader_stratified_2d(
+                    ncpaprop_atmosphere_reader_stratified_2d&&
+                        source ) noexcept :
+                    ncpaprop_atmosphere_reader_stratified_2d() {
+                    ::swap( *this, source );
+                }
+
+                virtual ~ncpaprop_atmosphere_reader_stratified_2d() {}
+
+                ncpaprop_atmosphere_reader_stratified_2d& operator=(
+                    ncpaprop_atmosphere_reader_stratified_2d other ) {
+                    ::swap( *this, other );
+                    return *this;
+                }
+
+                friend void ::swap(
+                    ncpaprop_atmosphere_reader_stratified_2d& a,
+                    ncpaprop_atmosphere_reader_stratified_2d& b ) noexcept;
+
+                virtual std::unique_ptr<_abstract_atmosphere_reader_2d> clone()
+                    const override {
+                    return std::unique_ptr<_abstract_atmosphere_reader_2d>(
+                        new ncpaprop_atmosphere_reader_stratified_2d(
+                            *this ) );
+                }
 
                 virtual bool stratified() const override { return true; }
         };
